Synchronizing code in external SVN repository with internal Google code base.
Revision created by MOE tool push_codebase.
MOE_MIGRATION=

+++ java/test/src/com/google/appengine/tools/pipeline/UserGuideTest.java
@@ -1,4 +1,16 @@
-// Copyright 2011 Google Inc. All Rights Reserved.
+// Copyright 2011 Google Inc.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not
+// use this file except in compliance with the License. You may obtain a copy of
+// the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+// WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+// License for the specific language governing permissions and limitations under
+// the License.
 
 package com.google.appengine.tools.pipeline;
 
@@ -11,7 +23,7 @@ import junit.framework.TestCase;
 
 /**
  * Tests for the sample code in the User Guide
- *
+ * 
  * @author rudominer@google.com (Mitch Rudominer)
  */
 public class UserGuideTest extends TestCase {
@@ -24,14 +36,14 @@ public class UserGuideTest extends TestCase {
     taskQueueConfig.setDisableAutoTaskExecution(false);
     taskQueueConfig.setShouldCopyApiProxyEnvironment(true);
     helper = new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig(), taskQueueConfig);
-
   }
 
   @Override
   public void setUp() throws Exception {
     super.setUp();
     helper.setUp();
-    System.setProperty("com.google.appengine.api.pipeline.use-simple-guids-for-debugging", "true");
+    System
+        .setProperty("com.google.appengine.api.pipeline.use-simple-guids-for-debugging", "true");
   }
 
   @Override
@@ -50,10 +62,10 @@ public class UserGuideTest extends TestCase {
     String pipelineId = service.startNewPipeline(new ComplexJob(), x, y, z);
     JobInfo jobInfo = service.getJobInfo(pipelineId);
     JobInfo.State state = jobInfo.getJobState();
-    if (JobInfo.State.COMPLETED_SUCCESSFULLY == state){
+    if (JobInfo.State.COMPLETED_SUCCESSFULLY == state) {
       System.out.println("The output is " + jobInfo.getOutput());
     }
-    int output = waitForJobToComplete(pipelineId);
+    int output = (Integer) waitForJobToComplete(pipelineId);
     assertEquals(((x - y) * (x - z)) - 2, output);
   }
